+++ utils/ft_putlptr.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putlptr.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: athi <athi@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: atkaewse <atkaewse@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/31 11:02:48 by athi              #+#    #+#             */
-/*   Updated: 2024/09/01 14:55:33 by athi             ###   ########.fr       */
+/*   Updated: 2024/09/01 15:10:27 by atkaewse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
 
-int	ft_putlptr(uintptr_t ptr)
+int	ft_putlptr(unsigned long ptr)
 {
 	size_t	len;
